Move tracing functions to TraceUtils
